% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gom.R
\name{calc_gom}
\alias{calc_gom}
\title{Title gom function}
\usage{
calc_gom(
  gom,
  max.iter = 1000,
  start_vec = c(0.520308489128769, 0.423437395856997, 1e-16, 0.0562541150142337),
  tol.gom = 1e-12,
  tol.lik = 1e-06
)
}
\arguments{
\item{gom}{: gom object}

\item{max.iter}{: max interaction}

\item{start_vec}{: starting}

\item{tol.gom}{: convergence}

\item{tol.lik}{: likelihood}
}
\value{
newlist
}
\description{
Title gom function
}
