% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gom.restart.R
\name{calc_gom.restart}
\alias{calc_gom.restart}
\title{Title GOM calcuation with restat in case no-convergence.}
\usage{
calc_gom.restart(
  gom,
  max.iter = 1000,
  start_vec = c(0.520308489128769, 0.423437395856997, 1e-16, 0.0562541150142337),
  tol.gom = 1e-12,
  tol.lik = 1e-06
)
}
\arguments{
\item{gom}{: gom object (pre-exsimated gom object)}

\item{max.iter}{: maximum interation}

\item{start_vec}{: starting value}

\item{tol.gom}{: convergence criteria}

\item{tol.lik}{: bound of likelihood}
}
\value{
a vector of c(updated_gom,iteration,restart.num)
}
\description{
Function that calculates gom scores. Input is a dataset with variables names, lambdas, and test case values
071717- added cond_func function that calculates conditional survival probabilities and life expectancies
}
