% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_func.R
\name{cond_func}
\alias{cond_func}
\title{Title Compute conditional survival functions}
\usage{
cond_func(GOM1, GOM2, GOM3, GOM4, tle_lam, dle_lam)
}
\arguments{
\item{dle_lam}{}
}
\value{
cond_func : returns
}
\description{
Title Compute conditional survival functions
}
