% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dle_nle.R
\name{dle_nle}
\alias{dle_nle}
\title{Title  Computing DFLE and DLE}
\usage{
dle_nle(GOM1, GOM2, GOM3, GOM4, tle_lam, dle_lam)
}
\arguments{
\item{dle_lam}{}
}
\value{
list(results, S_dle, S_nle)
}
\description{
Title  Computing DFLE and DLE
}
