% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gom.qing.R
\name{gom.qing}
\alias{gom.qing}
\title{Title Com computation based on the input data and gom object.}
\usage{
gom.qing(
  inputdata,
  inputdata_na,
  gom.obj,
  out.excel.file,
  start_vec1,
  start_vec2,
  max.iter = 1000,
  tol.gom = 10^(-12),
  tol.lik = 10^(-6)
)
}
\arguments{
\item{tol.lik}{}
}
\description{
Title Com computation based on the input data and gom object.
}
